Enabling OpenVINO: added configure `openvino`
Yaml config files under "cfg/examples/replication" were updated.
Config "openvino" added to specify the target device for an OpenVINO
model to execute.

So far, it is supported to execute GQCNN OpenVINO models on CPU, GPU,
or MYRIAD (Movidius NCS2).

This config is set to OFF by default, for GQCNN with Tensorflow
backend.

Signed-off-by: Sharron LIU <sharron.xr.liu@gmail.com>